+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,44 @@
 # e2e
 
+## 1.0.11
+
+### Patch Changes
+
+- feat: finalize transactions
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.39
+  - @midl-xyz/midl-js-react@0.0.35
+  - @midl-xyz/midl-js-core@0.0.29
+
+## 1.0.10
+
+### Patch Changes
+
+- 72b6eaa: feat: update builder
+- Updated dependencies [72b6eaa]
+  - @midl-xyz/midl-js-executor@0.0.38
+  - @midl-xyz/midl-js-react@0.0.34
+  - @midl-xyz/midl-js-core@0.0.28
+
+## 1.0.9
+
+### Patch Changes
+
+- fix: network detection
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.37
+  - @midl-xyz/midl-js-react@0.0.33
+  - @midl-xyz/midl-js-core@0.0.27
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.36
+  - @midl-xyz/midl-js-core@0.0.26
+  - @midl-xyz/midl-js-react@0.0.32
+
 ## 1.0.7
 
 ### Patch Changes